feat(masthead): transfer WIP to main masthead (#9213)
