--- packages/web-components/src/components/masthead/cloud/defs.ts
@@ -1,13 +1,0 @@
-/**
- * @license
- *
- * Copyright IBM Corp. 2022
- *
- * This source code is licensed under the Apache-2.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-export enum DDSMegamenuLayout {
-  TABBED = 'tabbed',
-  LISTING = 'listing',
-}
